requirements.txt
# Core ML/NLP libraries
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0

# Web scraping and APIs
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.8

# Data processing
nltk>=3.7
spacy>=3.4.0

# Web framework
fastapi>=0.95.0
uvicorn>=0.18.0

# Configuration and utilities
pydantic>=1.10.0
python-dotenv>=0.20.0
pyyaml>=6.0
click>=8.0.0

# Evaluation and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Database (optional)
sqlalchemy>=1.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Additional utilities
tqdm>=4.64.0
loguru>=0.6.0
pathlib2>=2.3.7

# RSS and enhanced parsing
feedparser>=6.0.8
lxml>=4.9.0
python-dateutil>=2.8.0

# Speed Optimization Dependencies
onnx>=1.12.0
onnxruntime>=1.12.0
onnxruntime-tools>=1.7.0

# Optional: For GPU acceleration
# torch-audio>=0.9.0  # Uncomment if using audio features
# torchvision>=0.10.0  # Uncomment if using image features